Refactor layout into reusable components
